# Eaas/openapi.yaml
openapi: 3.0.3
info:
  title: EaaS — Escrow as a Service API
  description: |
    API-first, Africa-native escrow platform. Enables secure transactions
    between buyers and sellers by holding funds in escrow until delivery is confirmed.
    
    ## Authentication
    - **Customer/Admin:** Bearer JWT token via `Authorization: Bearer <token>`
    - **Merchant API:** API Key via `X-API-Key: sk_live_xxxx` header
    - **Webhooks from PSP:** Verified via `X-Interswitch-Signature` HMAC header
    
    ## Idempotency
    All POST requests that create resources must include `X-Idempotency-Key` header.
    
    ## Rate Limiting
    - 100 requests/minute per API key
    - Rate limit headers returned: `X-RateLimit-Limit`, `X-RateLimit-Remaining`
    
  version: 1.0.0
  contact:
    name: EaaS Developer Support
    email: developers@eaas.africa
servers:
  - url: https://api.eaas.africa/api/v1
    description: Production
  - url: https://sandbox.eaas.africa/api/v1
    description: Sandbox
...
paths:
...
  /escrow/{reference}:
    get:
      tags: [Escrow]
      summary: Get escrow transaction by reference
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: path
          name: reference
          required: true
          schema:
            type: string
          example: "TXN-2024-A3F9K2"
      responses:
        '200':
          description: Escrow transaction details
        '404':
          $ref: '#/components/responses/NotFound'
...
  /disputes:
    post:
      tags: [Disputes]
      summary: Create a dispute
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDisputeRequest'
      responses:
        '201':
          description: Dispute created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /disputes/{id}:
    get:
      tags: [Disputes]
      summary: Get dispute details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dispute details
  /disputes/{id}/evidence:
    post:
      tags: [Disputes]
      summary: Upload evidence for a dispute
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Max 10MB. Accepted types — image, video, PDF.
                description:
                  type: string
      responses:
        '201':
          description: Evidence uploaded
        '413':
          description: File exceeds 10MB limit
        '400':
          description: Unsupported file type
...
  /admin/escrows:
    get:
      tags: [Admin]
      summary: List all escrow transactions (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: merchantId
          schema:
            type: string
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
        - in: query
          name: minAmount
          schema:
            type: number
        - in: query
          name: maxAmount
          schema:
            type: number
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paginated list
  /admin/disputes:
    get:
      tags: [Admin]
      summary: List all disputes (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [OPEN, UNDER_REVIEW, RESOLVED_MERCHANT, RESOLVED_CUSTOMER]
      responses:
        '200':
          description: Dispute list
  /admin/disputes/{id}/resolve:
    put:
      tags: [Admin]
      summary: Resolve a dispute
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution, resolutionNote]
              properties:
                resolution:
                  type: string
                  enum: [RESOLVED_MERCHANT, RESOLVED_CUSTOMER]
                resolutionNote:
                  type: string
                  minLength: 10
      responses:
        '200':
          description: Dispute resolved — payout or refund triggered
...
  /admin/fee-configurations:
    get:
      tags: [Admin]
      summary: List all fee configurations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fee configuration list
    post:
      tags: [Admin]
      summary: Create a fee configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeeConfigRequest'
      responses:
        '201':
          description: Fee configuration created
  /admin/fee-configurations/{id}:
    put:
      tags: [Admin]
      summary: Update a fee configuration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeeConfigRequest'
      responses:
        '200':
          description: Fee configuration updated
    delete:
      tags: [Admin]
      summary: Deactivate a fee configuration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fee configuration deactivated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          example:
            status: "error"
            message: "Merchant not found"
...
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          type: object
...
    CreateDisputeRequest:
      type: object
      required: [escrowReference, customerId, merchantId, raisedBy, reason, description, amountDisputed]
      properties:
        escrowReference:
          type: string
        customerId:
          type: string
          format: uuid
        merchantId:
          type: string
          format: uuid
        raisedBy:
          type: string
          format: uuid
        reason:
          type: string
        description:
          type: string
        desiredOutcome:
          type: string
        amountDisputed:
          type: number
    CreateFeeConfigRequest:
      type: object
      required: [feeType, feeValue]
      properties:
        merchantId:
          type: string
          format: uuid
        feeType:
          type: string
          enum: [PERCENTAGE, FLAT, BLENDED]
        feeValue:
          type: number
        minFee:
          type: number
          default: 500
        maxFee:
          type: number
          default: 50000
        effectiveFrom:
          type: string
          format: date-time
        effectiveUntil:
          type: string
          format: date-time
...